+++ ALZ.AKS/akstest-t01/terraform/modules/region/terraform.tf
@@ -21,5 +21,9 @@ terraform {
       source  = "hashicorp/time"
       version = "~> 0.12"
     }
+    tls = {
+      source  = "hashicorp/tls"
+      version = "~> 4.0"
+    }
   }
 }